Fixing addresses and sizes types

+++ src/managers/structures/bnb_and_pag/memory_manager.h
@@ -11,10 +11,10 @@ struct memory_manager
     byte *physical_memory;
     pcb *processes;
     pcb current_process;
-    int size;
+    size_t size;
 };
 
-memory_manager new_memory_manager(int memory_size);
+memory_manager new_memory_manager(size_t memory_size);
 void change_process_memory_manager(memory_manager manager, process_t in_process);
 void end_process_memory_manager(memory_manager manager, process_t in_process);
 #endif
